requirements.txt
Django==2.0.1
djangorestframework==3.7.7
celery==4.1.0
gunicorn==19.7.1
